Update Carts Screen And View Model

+++ app/src/main/java/com/example/ecommerce/data/model/DataXX.kt
@@ -2,6 +2,8 @@ package com.example.ecommerce.data.model
 
 data class DataXX(
     val cart_items: List<CartItem>,
-    val sub_total: Int,
-    val total: Int
+//    val sub_total: Int,
+//    val total: Int
+    val sub_total: Double,
+    val total: Double
 )
