+++ app/src/main/java/com/example/ecommerce/data/model/DataXXXX.kt
@@ -2,6 +2,8 @@ package com.example.ecommerce.data.model
 
 data class DataXXXX(
     val cart: Cart,
-    val sub_total: Int,
-    val total: Int
+//    val sub_total: Int,
+//    val total: Int
+    val sub_total: Double,
+    val total: Double
 )
