+++ app/src/main/java/com/example/ecommerce/data/model/ProductXXXX.kt
@@ -6,6 +6,8 @@ data class ProductXXXX(
     val id: Int,
     val image: String,
     val name: String,
-    val old_price: Int,
-    val price: Int
+    //val old_price: Int,
+    //val price: Int
+    val old_price: Double,
+    val price: Double
 )
