+++ app/src/main/java/com/example/ecommerce/data/model/ProductXXXXX.kt
@@ -4,6 +4,8 @@ data class ProductXXXXX(
     val discount: Int,
     val id: Int,
     val image: String,
-    val old_price: Int,
-    val price: Int
+//    val old_price: Int,
+//    val price: Int
+    val old_price: Double,
+    val price: Double
 )
